re-factor image/template matching functions into new matching module
